+++ Firestore/Swift/Source/Codable/CodablePassThroughTypes.swift
@@ -32,6 +32,14 @@ struct FirestorePassthroughTypes: StructureCodingPassthroughTypeResolver {
       t is Timestamp ||
       t is FieldValue ||
       t is DocumentReference ||
-      t is VectorValue
+      t is VectorValue ||
+      t is MinKey ||
+      t is MaxKey ||
+      t is RegexValue ||
+      t is Int32Value ||
+      t is Decimal128Value ||
+      t is BSONObjectId ||
+      t is BSONTimestamp ||
+      t is BSONBinaryData
   }
 }